MFF-pytorch/requirements.txt
numpy==1.17.2
opencv-python==4.1.1.26
Pillow>=6.2.2
pretrainedmodels==0.7.4
python-dateutil==2.8.0
torch==1.4.0
torchvision==0.5.0
pytz==2019.2
six==1.12.0
psutil
matplotlib
easydict
geffnet
pandas
imageio
torchinfo
openpyxl
timm==0.4.9
torch-optimizer==0.1.0
scikit-learn==0.24.2

# NOW RUN
# pip uninstall pillow
# pip install pillow-simd==7.0.0.post3

# INSTALL APEX Automatic Mixed Precision
# Install the Python only version
# git clone https://github.com/NVIDIA/apex && cd apex && pip install -v --disable-pip-version-check --no-cache-dir ./ 
# Further details can be found here: https://github.com/NVIDIA/apex#quick-start